Resolve conflicts and cleanup
