+++ src/constants/columns.js
@@ -3,20 +3,20 @@ import React from 'react';
 
 export const LOGS_COLUMNS = [
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.jobId" />,
-    value: 'jobId',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.id" />,
+    value: 'id',
     disabled: false,
     selected: false,
   },
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.bulkOperationType" />,
-    value: 'bulkOperationType',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.operationType" />,
+    value: 'operationType',
     disabled: false,
     selected: false,
   },
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.recordType" />,
-    value: 'recordType',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.entityType" />,
+    value: 'entityType',
     disabled: false,
     selected: false,
   },
@@ -27,32 +27,32 @@ export const LOGS_COLUMNS = [
     selected: false,
   },
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.runBy" />,
-    value: 'runBy',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.userId" />,
+    value: 'userId',
     disabled: false,
     selected: false,
   },
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.startedRunning" />,
-    value: 'startedRunning',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.startTime" />,
+    value: 'startTime',
     disabled: false,
     selected: false,
   },
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.endedRunning" />,
-    value: 'endedRunning',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.endTime" />,
+    value: 'endTime',
     disabled: false,
     selected: false,
   },
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.numberOfRecords" />,
-    value: 'numberOfRecords',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.totalNumOfRecords" />,
+    value: 'totalNumOfRecords',
     disabled: false,
     selected: false,
   },
   {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.processed" />,
-    value: 'processed',
+    label: <FormattedMessage id="ui-bulk-edit.columns.logs.processedNumOfRecords" />,
+    value: 'processedNumOfRecords',
     disabled: false,
     selected: false,
   },
