--- src/constants/constants.js
@@ -1,653 +1,0 @@
-import { FormattedMessage } from 'react-intl';
-import React from 'react';
-
-export const SUPPORTED_FILE_EXTENSIONS = ['csv'];
-
-export const BULK_EDIT_IDENTIFIERS = 'BULK_EDIT_IDENTIFIERS';
-export const BULK_EDIT_QUERY = 'BULK_EDIT_QUERY';
-
-export const BULK_EDIT_BARCODE = 'BARCODE';
-
-export const BULK_EDIT_UPDATE = 'BULK_EDIT_UPDATE';
-
-export const USER_COLUMNS = [
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.active" />,
-    value: 'active',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.lastName" />,
-    value: 'lastName',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.firstName" />,
-    value: 'firstName',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.barcode" />,
-    value: 'barcode',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.patronGroup" />,
-    value: 'patronGroup',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.username" />,
-    value: 'username',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.email" />,
-    value: 'email',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.expirationDate" />,
-    value: 'expirationDate',
-    disabled: false,
-    selected: false,
-  },
-];
-
-export const INVENTORY_COLUMNS_BASE = [
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.barcode" />,
-    value: 'barcode',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.status" />,
-    value: 'status',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.effectiveLocation" />,
-    value: 'effectiveLocation',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.callNumber" />,
-    value: 'callNumber',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.hrid" />,
-    value: 'hrid',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.materialType" />,
-    value: 'materialType',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.permanentLoanType" />,
-    value: 'permanentLoanType',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.temporaryLoanType" />,
-    value: 'temporaryLoanType',
-    disabled: false,
-    selected: true,
-  },
-];
-
-export const INVENTORY_COLUMNS = [
-  ...INVENTORY_COLUMNS_BASE,
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.id" />,
-    value: 'id',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.formerIds" />,
-    value: 'formerIds',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.accessionNumber" />,
-    value: 'accessionNumber',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.permanentLocation" />,
-    value: 'permanentLocation',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.temporaryLocation" />,
-    value: 'temporaryLocation',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.copyNumber" />,
-    value: 'copyNumber',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.enumeration" />,
-    value: 'enumeration',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.chronology" />,
-    value: 'chronology',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.volume" />,
-    value: 'volume',
-    disabled: false,
-    selected: false,
-  },
-];
-
-export const HOLDINGS_COLUMNS_BASE = [
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsHRID" />,
-    value: 'hrid',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsPermanentLocation" />,
-    value: 'permanentLocation',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsTemporaryLocation" />,
-    value: 'temporaryLocation',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsCallNumberPrefix" />,
-    value: 'callNumberPrefix',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsCallNumber" />,
-    value: 'callNumber',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsCallNumberSuffix" />,
-    value: 'callNumberSuffix',
-    disabled: false,
-    selected: true,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsType" />,
-    value: 'holdingsType',
-    disabled: false,
-    selected: true,
-  },
-];
-
-export const HOLDINGS_COLUMNS = [
-  ...HOLDINGS_COLUMNS_BASE,
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsEffectiveLocation" />,
-    value: 'effectiveLocation',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsUUID" />,
-    value: 'id',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsSource" />,
-    value: 'source',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsDiscoverySuppress" />,
-    value: 'discoverySuppress',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.holdingsCallNumberType" />,
-    value: 'callNumberType',
-    disabled: false,
-    selected: false,
-  },
-];
-
-export const CRITERIA = {
-  IDENTIFIER: 'identifier',
-  QUERY: 'query',
-  LOGS: 'logs',
-};
-
-export const CAPABILITIES = {
-  ITEM: 'ITEMS',
-  USER: 'USERS',
-  HOLDINGS: 'HOLDINGS_RECORD',
-};
-
-export const CAPABILITIES_VALUE = {
-  ITEMS: 'item',
-  USERS: 'user',
-  HOLDINGS_RECORD: 'holdings',
-};
-
-export const CAPABILITES_PREVIEW = {
-  ITEMS: 'items',
-  USERS: 'users',
-  HOLDINGS_RECORD: 'holdings',
-};
-export const ACTIONS = {
-  REPLACE: 'REPLACE_WITH',
-  CLEAR: 'CLEAR_FIELD',
-  FIND: 'FIND',
-};
-
-export const OPTIONS = {
-  TEMPORARY_LOCATION: 'TEMPORARY_LOCATION',
-  PERMANENT_LOCATION: 'PERMANENT_LOCATION',
-  STATUS: 'STATUS',
-  EXPIRATION_DATE: 'EXPIRATION_DATE',
-  EMAIL_ADDRESS: 'EMAIL_ADDRESS',
-  PATRON_GROUP: 'PATRON_GROUP',
-  TEMPORARY_LOAN_TYPE: 'TEMPORARY_LOAN_TYPE',
-  PERMANENT_LOAN_TYPE: 'PERMANENT_LOAN_TYPE',
-  TEMPORARY_HOLDINGS_LOCATION: 'TEMPORARY_LOCATION',
-  PERMANENT_HOLDINGS_LOCATION: 'PERMANENT_LOCATION',
-};
-
-export const getPlaceholder = (formatMessage) => ({
-  value: '',
-  label: formatMessage({ id: 'ui-bulk-edit.actions.placeholder' }),
-  disabled: true,
-});
-
-export const getFindAction = (formatMessage) => ({
-  value: ACTIONS.FIND,
-  label: formatMessage({ id: 'ui-bulk-edit.actions.find' }),
-  disabled: false,
-});
-
-export const getReplaceAction = (formatMessage) => ({
-  value: ACTIONS.REPLACE,
-  label: formatMessage({ id: 'ui-bulk-edit.layer.action.replace' }),
-  disabled: false,
-});
-
-export const getClearAction = (formatMessage) => ({
-  value: ACTIONS.CLEAR,
-  label: formatMessage({ id: 'ui-bulk-edit.layer.action.clear' }),
-  disabled: false,
-});
-
-export const PATRON_ACTIONS = (formatMessage) => [
-  getReplaceAction(formatMessage),
-];
-
-export const EXPIRATION_ACTIONS = (formatMessage) => [
-  getReplaceAction(formatMessage),
-];
-
-export const EMAIL_ACTIONS_FIRST = (formatMessage) => [
-  getFindAction(formatMessage),
-];
-
-export const EMAIL_ACTIONS_SECOND = (formatMessage) => [
-  getReplaceAction(formatMessage),
-];
-
-export const BASE_ACTIONS = (formatMessage) => [
-  getPlaceholder(formatMessage),
-  getReplaceAction(formatMessage),
-  getClearAction(formatMessage),
-];
-
-export const USER_OPTIONS = (formatMessage) => [
-  {
-    value: '',
-    label: formatMessage({ id: 'ui-bulk-edit.options.placeholder' }),
-    disabled: true,
-  },
-  {
-    value: OPTIONS.EMAIL_ADDRESS,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.email' }),
-    disabled: false,
-  },
-  {
-    value: OPTIONS.EXPIRATION_DATE,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.expirationDate' }),
-    disabled: false,
-  },
-  {
-    value: OPTIONS.PATRON_GROUP,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.patronGroup' }),
-    disabled: false,
-  },
-];
-
-export const HOLDINGS_OPTIONS = (formatMessage) => [
-  {
-    value: '',
-    label: formatMessage({ id: 'ui-bulk-edit.options.placeholder' }),
-    disabled: true,
-  },
-  { value: OPTIONS.TEMPORARY_HOLDINGS_LOCATION,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.holdings.temporaryLocation' }),
-    disabled: false },
-  {
-    value: OPTIONS.PERMANENT_HOLDINGS_LOCATION,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.holdings.permanentLocation' }),
-    disabled: false,
-  },
-];
-
-export const ITEMS_OPTIONS = (formatMessage) => [
-  {
-    value: '',
-    label: formatMessage({ id: 'ui-bulk-edit.options.placeholder' }),
-    disabled: true,
-  },
-  {
-    value: OPTIONS.TEMPORARY_LOCATION,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.temporaryLocation' }),
-    disabled: false,
-  },
-  {
-    value: OPTIONS.PERMANENT_LOCATION,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.permanentLocation' }),
-    disabled: false,
-  },
-  {
-    value: OPTIONS.STATUS,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.statusLabel' }),
-    disabled: false,
-  },
-  {
-    value: OPTIONS.TEMPORARY_LOAN_TYPE,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.temporaryLoanTypeLabel' }),
-    disabled: false,
-  },
-  {
-    value: OPTIONS.PERMANENT_LOAN_TYPE,
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.permanentLoanType' }),
-    disabled: false,
-  },
-];
-
-export const ITEM_STATUS_OPTIONS = (formatMessage) => [
-  {
-    value: '',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.placeholder' }),
-    disabled: true,
-  },
-  {
-    value: 'Available',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.available' }),
-    disabled: false,
-  },
-  {
-    value: 'Withdrawn',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.withdrawn' }),
-    disabled: false,
-  },
-  {
-    value: 'Missing',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.missing' }),
-    disabled: false,
-  },
-  {
-    value: 'In process (non-requestable)',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.non-requestable' }),
-    disabled: false,
-  },
-  {
-    value: 'Intellectual item',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.intellectual' }),
-    disabled: false,
-  },
-  {
-    value: 'Long missing',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.longMissing' }),
-    disabled: false,
-  },
-  {
-    value: 'Restricted',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.restricted' }),
-    disabled: false,
-  },
-  {
-    value: 'Unavailable',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.unavailable' }),
-    disabled: false,
-  },
-  {
-    value: 'Unknown',
-    label: formatMessage({ id: 'ui-bulk-edit.layer.options.unknown' }),
-    disabled: false,
-  },
-];
-
-export const TYPE_OF_PROGRESS = {
-  INITIAL: 'initial',
-  PROCESSED: 'processed',
-};
-
-export const JOB_STATUSES = {
-  SUCCESSFUL: 'SUCCESSFUL',
-  FAILED: 'FAILED',
-};
-
-export const itemColumnInAppWidths = {
-  barcode: '170px',
-  status: '110px',
-  effectiveLocation: '160px',
-  materialType: '100px',
-  permanentLoanType: '120px',
-  temporaryLoanType: '120px',
-};
-
-export const userColumnInAppWidths = {
-  active: '110px',
-  lastName: '110px',
-  firstName: '160px',
-  barcode: '100px',
-  patronGroup: '120px',
-  username: '120px',
-  email: '120px',
-  expirationDate: '120px',
-};
-
-export const holdingsColumnInAppWidths = {
-  hrid: '170px',
-  permanentLocation: '160px',
-  temporaryLocation: '100px',
-  callNumberPrefix: '120px',
-  callNumber: '120px',
-  callNumberSuffix: '120px',
-  holdingsType: '120px',
-};
-
-export const controlTypes = {
-  SELECT: 'SELECT',
-  DATE: 'DATE',
-  INPUT: 'INPUT',
-};
-
-export const translationSuffix = {
-  [CAPABILITIES.USER]: '',
-  [CAPABILITIES.ITEM]: '.item',
-  [CAPABILITIES.HOLDINGS]: '.holdings',
-};
-
-export const FILTERS = {
-  STATUS: 'status',
-  CAPABILITY: 'entityType',
-  OPERATION_TYPE: 'operationType',
-  START_DATE: 'startTime',
-  END_DATE: 'endTime',
-};
-
-export const FILTER_OPTIONS = {
-  STATUS: [
-    {
-      value: 'NEW',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.new" />,
-    },
-    {
-      value: 'RETRIEVING_RECORDS',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.retrieving" />,
-    },
-    {
-      value: 'SAVING_RECORDS_LOCALLY',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.saving" />,
-    },
-    {
-      value: 'DATA_MODIFICATION',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.data" />,
-    },
-    {
-      value: 'REVIEW_CHANGES',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.reviewing" />,
-    },
-    {
-      value: 'COMPLETED',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.completed" />,
-    },
-    {
-      value: 'COMPLETED_WITH_ERRORS',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.completedWithErrors" />,
-    },
-    {
-      value: 'FAILED',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.failed" />,
-    },
-  ],
-  CAPABILITY: [
-    {
-      value: 'HOLDING',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.holdings" />,
-    },
-    {
-      value: 'ITEM',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.items" />,
-    },
-    {
-      value: 'USER',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.users" />,
-    },
-  ],
-  OPERATION_TYPE: [
-    {
-      value: 'UPDATE',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.edit" />,
-    },
-    {
-      value: 'DELETE',
-      label: <FormattedMessage id="ui-bulk-edit.logs.filter.option.delete" />,
-    },
-  ],
-};
-
-export const LOGS_COLUMNS = [
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.id" />,
-    value: 'id',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.operationType" />,
-    value: 'operationType',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.entityType" />,
-    value: 'entityType',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.status" />,
-    value: 'status',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.userId" />,
-    value: 'userId',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.startTime" />,
-    value: 'startTime',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.endTime" />,
-    value: 'endTime',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.totalNumOfRecords" />,
-    value: 'totalNumOfRecords',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.processedNumOfRecords" />,
-    value: 'processedNumOfRecords',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.editing" />,
-    value: 'editing',
-    disabled: false,
-    selected: false,
-  },
-  {
-    label: <FormattedMessage id="ui-bulk-edit.columns.logs.actions" />,
-    value: 'actions',
-    disabled: false,
-    selected: false,
-  },
-];
