+++ src/utills/formatters.js
@@ -2,17 +2,18 @@ import { InfoPopover } from '@folio/stripes/components';
 import { FormattedMessage } from 'react-intl';
 import React from 'react';
 import BulkEditLogsActions from '../components/BulkEditLogs/BulkEditLogsActions/BulkEditLogsActions';
+import { FormattedTime } from '../utils/FormattedTime';
 
-export const getLogsResultsFormatter = () => ({
-  jobId: item => item.jobId,
-  bulkOperationType: item => item.bulkOperationType,
-  recordType: item => item.recordType,
+export const getLogsResultsFormatter = (userNamesMap) => ({
+  id: item => item.id,
+  operationType: item => item.operationType,
+  entityType: item => item.entityType,
   status: item => item.status,
-  runBy: item => item.runBy,
-  startedRunning: item => item.startedRunning,
-  endedRunning: item => item.endedRunning,
-  numberOfRecords: item => item.numberOfRecords,
-  processed: item => item.processed,
+  userId: item => userNamesMap[item.userId],
+  startTime: item => <FormattedTime dateString={item.startTime} />,
+  endTime: item => <FormattedTime dateString={item.endTime} />,
+  totalNumOfRecords: item => item.totalNumOfRecords,
+  processedNumOfRecords: item => item.processedNumOfRecords,
   editing: item => item.editing,
   actions: (item) => (
     <>
